+++ frontend/src/components/IndividualModel/CDFTable.js
