+++ frontend/src/components/IndividualModel/ContinuousDeviance.js
