+++ frontend/src/components/IndividualModel/ContinuousSummary.js
