+++ frontend/src/components/IndividualModel/DichotomousDeviance.js
