+++ frontend/src/components/IndividualModel/DichotomousSummary.js
