+++ frontend/src/components/IndividualModel/GoodnessFit.js
